requirements.txt
markdown>=3.4.0
bleach>=5.0.0
python-magic>=0.4.0
Pillow>=9.0.0
python-dotenv>=0.19.0
Flask-WTF>=1.0.0
